+++ home.admin/config.scripts/cl.install.sh
@@ -2,7 +2,7 @@
 # https://lightning.readthedocs.io/
 
 # https://github.com/ElementsProject/lightning/releases
-CLVERSION="v23.05.2"
+CLVERSION="v23.08"
 
 # install the latest master by using the last commit id
 # https://github.com/ElementsProject/lightning/commit/master
@@ -58,7 +58,7 @@ function installDependencies() {
 function buildAndInstallCLbinaries() {
   echo "- Configuring EXPERIMENTAL_FEATURES enabled"
   echo
-  sudo -u bitcoin ./configure --enable-experimental-features
+  sudo -u bitcoin ./configure
   echo
   echo "- Building Core lightning from source"
   echo
@@ -367,9 +367,6 @@ alias ${netprefix}clconf=\"sudo nano ${CLCONF}\"
   # setting values in the raspiblitz.conf
   /home/admin/config.scripts/blitz.conf.sh set ${netprefix}cl on
   # blitz.conf.sh needs sudo access - cannot be run in cl.check.sh
-  if [ ! -f /home/bitcoin/${netprefix}cl-plugins-enabled/sparko ]; then
-    /home/admin/config.scripts/blitz.conf.sh set ${netprefix}sparko "off"
-  fi
   if [ ! -f /home/bitcoin/cl-plugins-enabled/c-lightning-http-plugin ]; then
     /home/admin/config.scripts/blitz.conf.sh set clHTTPplugin "off"
   fi